componentizes Joystick from Sequencer -> Channel
